requirements.txt
asgiref==3.11.1
Django==6.0.2
django-debug-toolbar==6.2.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
PyJWT==2.11.0
sqlparse==0.5.5
